Significant refactoring plus some additions
Refactored the code to be more object oriented to support potential future additions. Created classes to represent the game state, GUI, and database, and encapsulated most functions in these classes.

Added a column to the leaderboard to show which word was guessed.

Fixed the menu buttons such that viewing the scoreboard doesn't disrupt a game in progress.